requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file requirements.txt requirements.in
#
Click==6.7
logutils==0.3.4.1
MySQL-python==1.2.5       # via mysql
mysql==0.0.1
